Implemented rudimentary timezone support
There is an interesting issue currently where using UTC can cause
the desired time range to cross midnight, such that doing a simple
range. A simple work-around is to use timezones, although another
option could be to switch to a start-time and length instead of
a start-time, stop-time.

Currently, however, a timezone modifier can be used to use localtime
instead.