Fixed Arduino freezing due to NTP timeout
It seems that when NTP would be unavailable, the Arduino would
freeze. I haven't figured out if it is due to a UDP issue or
an NTP one. As a current work-around, now time is only synced
at boot and when pressing the down button on the LCD shield.
That way, the network being unavailable shouldn't impact the
Arduino.